Massive updates. Check the description
Now using Python Shelves in order to persist the card dictionary throughout the run.
Now only fires off notifications if the card is actually in stock.
Offloaded the HTML parsing to the Card model class.
Changed the back-off to a random number between 10 and 30 seconds.  I'm still trying to figure out a better way to do this, let me know (or pull) if you can think of one!
Cleaned up variable and function names to make them more Pythonic.